Fully working exponential back-off message polling.
Message polling is now done in a separate  dedicated thread instead of using an intent service, hence the service java file has been deleted and its references across the application and the manifest file as-well.

Polling starts and ends just as before, with main activity active status (between onResume and onPause).

The repeatable action running that depends on some sort of backoff mechanism was written in its own class RepeatedBackoffTaskRunner.
It has its own implementation that instantiate it in a way that it uses a dedicated  thread (it is possible to make others TaskRunners that would use the same thread)
It uses an exponential backoff strategy and does the polling action via the IMessagePoller, made earlier (way-way earlier....)
The packages hierarchy changed a bit to fit the new state where no service is used for the message polling.

Auto-indentation across all "touched" files